# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Plant Doctor API",
        "description": "API for plant disease diagnosis and health tracking",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8005",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/api/plants": {
            "get": {
                "tags": [
                    "Plants"
                ],
                "summary": "List user plants",
                "description": "GET /api/plants\nList user plants",
                "operationId": "6e2a3a91967b1e364c87b755cd5152f6",
                "responses": {
                    "200": {
                        "description": "Plants listed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Plant"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Plants"
                ],
                "summary": "Create a new plant",
                "description": "POST /api/plants\nCreate a new plant",
                "operationId": "d005e00370ab9bd321d50ba5dde92042",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/StorePlantRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Plant created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Plant"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
        "/api/plants/{plant}": {
            "get": {
                "tags": [
                    "Plants"
                ],
                "summary": "Get plant details",
                "description": "GET /api/plants/{plant}\nGet plant details",
                "operationId": "6ce214db46790bc4016abfac89457334",
                "parameters": [
                    {
                        "name": "plant",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Plant retrieved",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Plant"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Plants"
                ],
                "summary": "Update a plant",
                "description": "PUT /api/plants/{plant}\nUpdate a plant",
                "operationId": "456d95602efa0a2cafadd6f37ce28dee",
                "parameters": [
                    {
                        "name": "plant",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdatePlantRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Plant updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Plant"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Plants"
                ],
                "summary": "Delete a plant",
                "description": "DELETE /api/plants/{plant}\nDelete a plant",
                "operationId": "81eb3303dd0676cd7cf28ae887e7cbda",
                "parameters": [
                    {
                        "name": "plant",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Plant deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "Diagnosis": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "confidence_score": {
                        "type": "number",
                        "format": "float",
                        "example": 87.5
                    },
                    "notes": {
                        "type": "string",
                        "nullable": true
                    },
                    "image_url": {
                        "type": "string",
                        "format": "uri",
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "example": "completed"
                    },
                    "ai_provider": {
                        "description": "AI service used for diagnosis",
                        "type": "string",
                        "example": "groq",
                        "nullable": true
                    },
                    "species": {
                        "properties": {
                            "name": {
                                "description": "Scientific name of the identified species",
                                "type": "string",
                                "example": "Rosa canina",
                                "nullable": true
                            },
                            "common_names": {
                                "description": "Common names of the species",
                                "type": "array",
                                "items": {
                                    "type": "string"
                                },
                                "example": [
                                    "dog rose",
                                    "escaramujo"
                                ]
                            }
                        },
                        "type": "object"
                    },
                    "disease": {
                        "properties": {
                            "name": {
                                "description": "Common name of the detected disease",
                                "type": "string",
                                "example": "Black Spot",
                                "nullable": true
                            },
                            "scientific_name": {
                                "description": "Scientific name of the disease",
                                "type": "string",
                                "example": "Diplocarpon rosae",
                                "nullable": true
                            },
                            "severity": {
                                "type": "string",
                                "example": "medium",
                                "nullable": true,
                                "enum": [
                                    "low",
                                    "medium",
                                    "high"
                                ]
                            },
                            "symptoms": {
                                "description": "Observed symptoms",
                                "type": "array",
                                "items": {
                                    "type": "string"
                                },
                                "example": [
                                    "Dark circular spots",
                                    "Yellowing leaves"
                                ]
                            },
                            "treatment": {
                                "description": "Recommended treatment",
                                "type": "string",
                                "nullable": true
                            },
                            "prevention": {
                                "description": "Prevention measures",
                                "type": "string",
                                "nullable": true
                            },
                            "catalog_match": {
                                "$ref": "#/components/schemas/Disease"
                            }
                        },
                        "type": "object"
                    },
                    "expert_verified": {
                        "type": "boolean",
                        "example": false
                    },
                    "expert_notes": {
                        "type": "string",
                        "nullable": true
                    },
                    "plant": {
                        "$ref": "#/components/schemas/Plant"
                    },
                    "expert": {
                        "$ref": "#/components/schemas/User"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Disease": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Oídio"
                    },
                    "scientific_name": {
                        "type": "string",
                        "example": "Erysiphe cichoracearum",
                        "nullable": true
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "symptoms": {
                        "type": "string",
                        "nullable": true
                    },
                    "treatment": {
                        "type": "string",
                        "nullable": true
                    },
                    "prevention": {
                        "type": "string",
                        "nullable": true
                    },
                    "image_url": {
                        "type": "string",
                        "format": "uri",
                        "nullable": true
                    },
                    "category": {
                        "type": "string",
                        "example": "fungal",
                        "nullable": true
                    },
                    "severity": {
                        "type": "string",
                        "example": "medium"
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "perenual_id": {
                        "description": "Perenual API ID",
                        "type": "integer",
                        "nullable": true
                    },
                    "last_synced_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "ErrorResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Error occurred"
                    }
                },
                "type": "object"
            },
            "Plant": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Mi Tomate"
                    },
                    "species": {
                        "type": "string",
                        "example": "Solanum lycopersicum",
                        "nullable": true
                    },
                    "description": {
                        "type": "string",
                        "example": "Tomate del jardín",
                        "nullable": true
                    },
                    "image_url": {
                        "type": "string",
                        "format": "uri",
                        "nullable": true
                    },
                    "latitude": {
                        "type": "number",
                        "format": "float",
                        "nullable": true
                    },
                    "longitude": {
                        "type": "number",
                        "format": "float",
                        "nullable": true
                    },
                    "country_name": {
                        "type": "string",
                        "nullable": true
                    },
                    "state_name": {
                        "type": "string",
                        "nullable": true
                    },
                    "city_name": {
                        "type": "string",
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "example": "active"
                    },
                    "diagnoses_count": {
                        "type": "integer",
                        "example": 3
                    },
                    "diagnoses": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Diagnosis"
                        }
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "StorePlantRequest": {
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Mi Tomate"
                    },
                    "species": {
                        "type": "string",
                        "example": "Solanum lycopersicum",
                        "nullable": true
                    },
                    "description": {
                        "type": "string",
                        "example": "Tomate del jardín",
                        "nullable": true
                    },
                    "image": {
                        "type": "string",
                        "format": "binary",
                        "nullable": true
                    },
                    "latitude": {
                        "type": "number",
                        "format": "float",
                        "example": 19.4326,
                        "nullable": true
                    },
                    "longitude": {
                        "type": "number",
                        "format": "float",
                        "example": -99.1332,
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "UpdatePlantRequest": {
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Mi Tomate Actualizado"
                    },
                    "species": {
                        "type": "string",
                        "example": "Solanum lycopersicum",
                        "nullable": true
                    },
                    "description": {
                        "type": "string",
                        "example": "Tomate actualizado",
                        "nullable": true
                    },
                    "image": {
                        "type": "string",
                        "format": "binary",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "SuccessResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Operation successful"
                    },
                    "data": {
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "ValidationErrorResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Validation failed"
                    },
                    "errors": {
                        "type": "object",
                        "example": {
                            "email": [
                                "The email field is required."
                            ]
                        }
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "jwt": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}